Remove uneccessary aria
