+++ packages/spor-react/src/datepicker/DateTimeSegment.tsx
@@ -2,12 +2,12 @@ import { Box, useMultiStyleConfig } from "@chakra-ui/react";
 import React, { RefObject, forwardRef, useRef } from "react";
 import { useDateSegment } from "react-aria";
 import { DateFieldState, DateSegment } from "react-stately";
-import { createTexts, useTranslation } from "../i18n";
 
 type DateTimeSegmentProps = {
   segment: DateSegment;
   state: DateFieldState;
-  ariaLabelledby?: string;
+  ariaLabel?: string;
+  ariaDescription?: string;
 };
 /**
  * A date time segment is a part of a date or a time stamp.
@@ -17,11 +17,9 @@ type DateTimeSegmentProps = {
  * This component should be used with the react-aria library, and is not meant to be used directly.
  * */
 export const DateTimeSegment = forwardRef<HTMLDivElement, DateTimeSegmentProps>(
-  ({ segment, state, ariaLabelledby }, externalRef) => {
+  ({ segment, state, ariaLabel, ariaDescription }, externalRef) => {
     const internalRef = useRef(null);
     const ref = externalRef ?? internalRef;
-
-    const { t } = useTranslation();
 
     const { segmentProps } = useDateSegment(
       segment,
@@ -46,8 +44,8 @@ export const DateTimeSegment = forwardRef<HTMLDivElement, DateTimeSegmentProps>(
         borderRadius="xs"
         fontSize={["mobile.sm", "desktop.sm"]}
         sx={styles.dateTimeSegment}
-        aria-labelledby={ariaLabelledby}
-        aria-label={t(getAriaLabel(segment.type))}
+        aria-description={ariaDescription}
+        aria-labelledby={ariaLabel}
       >
         {isPaddable(segment.type)
           ? segment.text.padStart(2, "0")
@@ -63,37 +61,3 @@ const isPaddable = (segmentType: DateSegment["type"]) =>
   segmentType === "hour" ||
   segmentType === "minute" ||
   segmentType === "second";
-
-const texts = createTexts({
-  day: {
-    nb: "Velg dag",
-    nn: "Vel dag",
-    sv: "Välj dag",
-    en: "Choose day",
-  },
-  month: {
-    nb: "Velg måned",
-    nn: "Vel månad",
-    sv: "Välj månad",
-    en: "Choose month",
-  },
-  year: {
-    nb: "Velg år",
-    nn: "Vel år",
-    sv: "Välj år",
-    en: "Choose year",
-  },
-});
-
-const getAriaLabel = (segmentType: DateSegment["type"]) => {
-  switch (segmentType) {
-    case "day":
-      return texts.day;
-    case "month":
-      return texts.month;
-    case "year":
-      return texts.year;
-    default:
-      return texts.day;
-  }
-};
